Move logic to service and test service instead of cronjob.
Refactor to using result type in service
